>Err: Permission Denied
Only i put sudo (super user do) before the mv devsuite /usr/bin/ 
Because if the user execute the command that I put in README.md sudo chmod a+x install.sh && ./install.sh it says that the permission is denied to move to a only-root dir, the other solution may is to put ... && sudo ./install.sh. But also i'm the chief of this open-source project and decide to do that.